ENHANCED: If --on-warning=status or --on-error=status is active, consider such messages testv failures
When running tests using ``--on-error=status``, printing an error message that
does not cause a test to fail silently ignores the message while Prolog
exits with status 1.  That is hard to debug.   Now, if these flags are
present, tests that emit warnings or errors are considered failed and
thus the test is printed.